git-svn-id: http://ironman.agnux.com/repositorios/ErpAgnuxV1@913 ce483f98-b20a-4c78-8f99-c15376fc4320

+++ erp-quimicos/src/main/java/com/agnux/kemikal/controllers/RepVentasNetasProductoFacturaController.java
@@ -452,11 +452,12 @@ public class RepVentasNetasProductoFacturaController {
                         registros_normales.put("total_venta", String.valueOf(0.0));
 
                         sumatoria_cantidad = sumatoria_cantidad + Double.parseDouble(registro.get("cantidad"));
-                        if (Integer.parseInt(registro.get("costo"))== 0){
+                        if (Double.parseDouble(registro.get("costo"))== 0){
                              costo=0.0;
                         }else{
                         costo=Double.parseDouble(registro.get("costo"));
                         }
+                        
                         sumatoria_costo = sumatoria_costo + costo;
                         
                         sumatoria_venta = sumatoria_venta + Double.parseDouble(registro.get("venta_pesos"));
@@ -530,7 +531,7 @@ public class RepVentasNetasProductoFacturaController {
 
                         //sumar primera cantidad del nuevo cliente
                         sumatoria_cantidad = sumatoria_cantidad + Double.parseDouble(registro.get("cantidad"));
-                        if (Integer.parseInt(registro.get("costo"))== 0){
+                        if (Double.parseDouble(registro.get("costo"))== 0){
                              costo=0.0;
                         }else{
                         costo=Double.parseDouble(registro.get("costo"));
